# src/zk/blockbuster/swagger.yml
openapi: 3.0.0
info:
  title: Blockbuster
  version: 0.1.0
  description: A super lightweight directory service with lease management.
...
paths:
...
  /lease:
    get:
      tags:
        - lease
      summary: Get Leases.
      operationId: getLeases
      responses:
        200:
          description: Lease list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lease'
    post:
      tags:
        - lease
      summary: Create a lease from specific configuration items.
      operationId: createLease
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                duration:
                  type: integer
                  example: 60
                  description: lease duration in seconds
                uri:
                  $ref: '#/components/schemas/Uri'
                scope:
                  $ref: '#/components/schemas/SearchScope'
      responses:
        200:
          description: Success to create a lease.
          content:
            application/json:
              schema:
                result:
                  type: boolean
        404:
          description: CI not found.
        400:
          description: param is invalid.

    delete:
      tags:
        - lease
      summary: delete a lease
      operationId: deleteLease
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                uri:
                  $ref: '#/components/schemas/Uri'
      responses:
        200:
          description: Success to create a lease.
          content:
            application/json:
              schema:
                result:
                  type: boolean
        404:
          description: CI not found.
        400:
          description: param is invalid.

    put:
      tags:
        - lease
      summary: update a lease
      operationId: updateLease
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                uri:
                  $ref: '#/components/schemas/Uri'
                renew:
                  type: integer
                  example: 60
                  description: Renew period. The updated lease will expire after current time + renew period.
      responses:
        200:
          description: Success to create a lease.
          content:
            application/json:
              schema:
                result:
                  type: boolean
        404:
          description: CI not found.
        400:
          description: param is invalid.
components:
  schemas:
    Uri:
      type: string
      example: /p1/p2/p3
...
    SearchScope:
      type: integer
      enum: [0, 1, 2]
      default: 0
      description: 0 for base object, 1 for one-level, 2 for subtree (TBD)
      externalDocs:
        description: LDAP's convention
        url: https://docs.microsoft.com/en-us/windows/win32/ad/search-scope
    ConfigurationItem:
      type: object
      properties:
        id:
          type: string
          readOnly: true
          format: uuid
        uri:
          type: string
          example: /p1/p2/p3
        data:
          type: object
          example:
            field1: value1
        maxLease:
          type: integer
          default: 1
          example: 1
          description: max lease a CI can have
    Lease:
      type: object
      properties:
        id:
          type: string
          readOnly: true
          format: uuid
        expiredAfter:
          type: string
          format: date-time
        ci:
          $ref: '#/components/schemas/ConfigurationItem'